+++ src/SQLProvider/Providers.MSAccess.fs
@@ -317,8 +317,8 @@ type internal MSAccessProvider() =
                         | false -> sprintf "[%s].[%s]" al col
                     let fieldNotationAlias(al:alias,col:string) =
                         match String.IsNullOrEmpty(al) with
-                        | true -> sprintf "[%s]" col
-                        | false -> sprintf "[%s_%s]" al col
+                        | true -> sprintf "'[%s]'" col
+                        | false -> sprintf "'[%s][%s]'" al col
 
                     match sqlQuery.Grouping with
                     | [] -> FSharp.Data.Sql.Common.Utilities.parseAggregates fieldNotation fieldNotationAlias sqlQuery.AggregateOp