+++ src/SQLProvider/Providers.MsSqlServer.fs
@@ -604,8 +604,8 @@ type internal MSSqlServerProvider(tableNames:string) =
                         | false -> sprintf "[%s].[%s]" al col
                     let fieldNotationAlias(al:alias,col:string) = 
                         match String.IsNullOrEmpty(al) with
-                        | true -> sprintf "'%s'" col
-                        | false -> sprintf "'[%s%s]'" al col
+                        | true -> sprintf "'[%s]'" col
+                        | false -> sprintf "'[%s][%s]'" al col
 
                     match sqlQuery.Grouping with
                     | [] -> FSharp.Data.Sql.Common.Utilities.parseAggregates fieldNotation fieldNotationAlias sqlQuery.AggregateOp