+++ src/SQLProvider/Providers.MySql.fs
@@ -552,8 +552,8 @@ type internal MySqlProvider(resolutionPath, owner, referencedAssemblies) as this
                         | false -> sprintf "`%s`.`%s`" al col
                     let fieldNotationAlias(al:alias,col:string) =
                         match String.IsNullOrEmpty(al) with
-                        | true -> sprintf "`%s`" col
-                        | false -> sprintf "'`%s`.`%s`'" al col
+                        | true -> sprintf "`[%s]`" col
+                        | false -> sprintf "`[%s][%s]`" al col
 
                     match sqlQuery.Grouping with
                     | [] -> FSharp.Data.Sql.Common.Utilities.parseAggregates fieldNotation fieldNotationAlias sqlQuery.AggregateOp