+++ src/SQLProvider/Providers.Oracle.fs
@@ -652,8 +652,8 @@ type internal OracleProvider(resolutionPath, owner, referencedAssemblies, tableN
                         | false -> sprintf "%s.%s" al col
                     let fieldNotationAlias(al:alias,col:string) =
                         match String.IsNullOrEmpty(al) with
-                        | true -> sprintf "\"%s\"" col
-                        | false -> sprintf "\"%s.%s\"" al col
+                        | true -> sprintf "\"[%s]\"" col
+                        | false -> sprintf "\"[%s][%s]\"" al col
 
                     match sqlQuery.Grouping with
                     | [] -> FSharp.Data.Sql.Common.Utilities.parseAggregates fieldNotation fieldNotationAlias sqlQuery.AggregateOp